feat(nika-lsp): the three 557 lanes — tools whitelist · task members · schema context
The remaining completion gaps the editor-side probe pinned (#557),
each a lane the server already had the truth for:

- an agent's tools: [ … ] whitelist speaks the catalog — the same
  register the singular tool: value gets (MCP refs stay the
  author's: the server has no MCP registry to speak from). The
  open-bracket detector is now SCOPED: an unclosed [ abandoned
  upstream dies at the first line back at the key's indent instead
  of capturing every later position (the probe caught the leak;
  in_open_depends_on inherits the same containment).
- ${{ tasks.<id>. }} teaches the task's member facts — output
  (verb-aware detail) · status · error · plus the task's own named
  output: bindings. Line-scanned, not parsed: the open island makes
  the document YAML-invalid at the very moment the lane fires, so a
  parse-first branch would be dead code.
- a schema: block speaks JSON Schema, never the task fields (the
  probe's bug: schema: offered depends_on). Direct children of
  schema:/items: get the keyset (type · required · properties ·
  items · enum · description); children of properties: stay silent
  — those names belong to the author. The task-field lane is
  guarded by the ancestor chain.

168 lib tests (10 new: the three lanes, the containment regression,
the properties silence, the CEL post-dot ordering). Probed live
over JSON-RPC on the built binary — the three payloads answer.
